requirements.txt
stripe
openai
flask
python-dotenv